+++ PDFUtils.h
@@ -66,13 +66,13 @@ void displayBox(PIX *pix, BOX *box);
 bool isFilledByImage(PDFDoc *doc, int page);
 
 // Gets a PIX of the given page rendered at the given dpi.
-std::unique_ptr<PIX> getFullRenderPix(PDFDoc *doc, int page, double dpi);
+std::unique_ptr<PIX> getFullRenderPix(PDFDoc *doc, int page, int dpi);
 
 // Gets a PIX of the given page rendered wthout text at the given dpi.
-std::unique_ptr<PIX> getGraphicOnlyPix(PDFDoc *doc, int page, double dpi);
+std::unique_ptr<PIX> getGraphicOnlyPix(PDFDoc *doc, int page, int dpi);
 
 // Gets the TextPage* objects of a document at a given dpi.
-std::vector<TextPage *> getTextPages(PDFDoc *doc, double dpi);
+std::vector<TextPage *> getTextPages(PDFDoc *doc, int dpi);
 
 // Get a PIX with the given figures drawn on it
 PIX *drawFigureRegions(PIX *background, const std::vector<Figure> &figures);
@@ -83,13 +83,13 @@ GooString *jsonSanitizeUTF8(GooString *str);
 
 void writeText(TextPage *page, BOX *bb, const char *name, std::ostream &output);
 
-void saveFiguresJSON(std::vector<Figure> &figures, PIX *original, double dpi,
+void saveFiguresJSON(std::vector<Figure> &figures, PIX *original, int dpi,
                      std::string prefix, std::vector<TextPage *> text);
 
 void saveFiguresImage(std::vector<Figure> &figures, PIX *original,
                       std::string prefix);
 
-void saveFigures(std::vector<Figure> &figures, PIX *original, double dpi,
+void saveFigures(std::vector<Figure> &figures, PIX *original, int dpi,
                  std::vector<TextPage *> pages, std::string imagePrefix,
                  std::string jsonPrefix);
